plivo_pub_sub/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
channels==4.1.0
daphne==4.1.0
websockets==12.0
